Change generated condition for resolver.
The generated conditions for resolvers and predicates are now equal.
First, the token kind is checked and the condition is then added.
Previously, only the condition was generated for a resolver. The
user has to remember that the token kind must be checked, too. This
is an unnecessary burden and an easy source for errors.