Change generation for alternatives.
If an alternative is the first element of an optional group, then
there is no need to check that one alternative is matched, because
this is already covered by the if/while from the optional group.

E.g. consider (A | B)? . The parser has to check if the next token
is in first set of A or B in order to enter the group. Inside the
group it is not necessary to check that the token is not in first
aet of A and B, because it's not possible.