Put parsing functionality into a block.
This prevents the issue that `goto _error;` could jump beyond an
initialized object. This gives more freedom in defining variables
in code blocks. One consequence is that the values are not available
in the _error block. At least the first code block at the top could
be placed outside this new block to allow some communication in case
it will become necessary.